+++ NEWS.md
@@ -1,3 +1,8 @@
+# plotdap 0.0.5
+* fixed problem with animation in 'add_griddap()'
+* fixed bad URL for sardine example
+* changed to use colors in package 'cmocean' instead of in 'rerddap'
+
 # plotdap 0.0.4
 * fixed changed SODA datasetid in vignette
 * fixed Namespace and Import problems